+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.5.0-rc.13",
+  sdkVersion: "4.5.0-rc.14",
   genVersion: "2.844.3",
-  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.13 2.844.3 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.14 2.844.3 2.0 @orq-ai/node",
 } as const;