+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -148,7 +148,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-26T12:31:53.721Z",
+    "2026-02-26T14:56:55.966Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {